requirements.txt
aiofiles
aiogram
aiohttp
aiosignal
annotated-types
attrs
emoji
certifi
frozenlist
httpx
idna
magic-filter
multidict
pydantic
pydantic-settings
pydantic_core
python-dotenv
typing_extensions
yarl
yadisk